Refactor integration tests: reorganize and enhance test structure
- Add new integration test files for ping repository and service, consolidating tests into domain-specific directories.
- Update existing tests to improve clarity and performance, including context handling and performance checks.
- Remove outdated integration test files to streamline the codebase.

These changes enhance maintainability and test discoverability by co-locating tests with their respective components.
