--- PR_DESCRIPTION.md
@@ -1,172 +1,0 @@
-# 🔄 Refactor: Reorganize Integration Tests by Domain
-
-## 📋 Summary
-
-This PR reorganizes integration tests by moving them from a centralized `integration_tests/` directory to domain-specific locations within the `internal/ping/` structure. This improves maintainability, test discoverability, and follows better testing practices.
-
-## 🎯 Motivation
-
-The previous structure had all integration tests in a centralized `integration_tests/` directory, which made it difficult to:
-- Find tests related to specific domains
-- Maintain tests alongside their code
-- Test specific layers independently
-- Understand the relationship between code and tests
-
-## 🔧 Changes Made
-
-### ✅ **New Test Structure**
-
-```
-internal/
-├── ping/
-│   ├── module_integration_test.go          # Complete module integration tests
-│   ├── adapter/
-│   │   ├── ping_repo/
-│   │   │   └── repository_integration_test.go  # Repository integration tests
-│   │   └── rest/
-│   │       └── ping_handler_integration_test.go # Handler integration tests
-│   └── core/
-│       └── service/
-│           └── service_integration_test.go      # Service integration tests
-```
-
-### ✅ **Files Added**
-
-- `internal/ping/module_integration_test.go` - Complete module end-to-end testing
-- `internal/ping/adapter/ping_repo/repository_integration_test.go` - Repository layer testing
-- `internal/ping/adapter/rest/ping_handler_integration_test.go` - Handler layer testing
-- `internal/ping/core/service/service_integration_test.go` - Service layer testing
-- `docs/INTEGRATION_TESTING.md` - Updated documentation
-- `scripts/integration_test_runner.sh` - Updated test runner script
-
-### ✅ **Files Modified**
-
-- `Makefile` - Updated test paths for integration tests
-- `scripts/integration_test_runner.sh` - Updated to use new test structure
-- `docs/INTEGRATION_TESTING.md` - Added documentation for new structure
-
-### ✅ **Files Removed**
-
-- `integration_tests/` directory and all its contents
-
-## 🧪 Test Categories
-
-Each integration test file includes comprehensive testing for:
-
-### **Module Integration Tests** (`module_integration_test.go`)
-- Complete end-to-end workflow testing
-- Performance testing
-- Error handling scenarios
-- Load testing
-- Database and cache connectivity (when available)
-
-### **Repository Integration Tests** (`repository_integration_test.go`)
-- Basic repository functionality
-- Performance testing
-- Context handling
-- Database connectivity (when available)
-- Cache connectivity (when available)
-- Error scenarios
-
-### **Service Integration Tests** (`service_integration_test.go`)
-- Service layer with real dependencies
-- Performance testing
-- Context handling
-- Database and cache integration (when available)
-- Error scenarios
-
-### **Handler Integration Tests** (`ping_handler_integration_test.go`)
-- HTTP endpoint health checks
-- Performance testing
-- Response headers validation
-- Error handling scenarios
-- Load testing
-
-## 🚀 Benefits
-
-### **Improved Maintainability**
-- Tests are co-located with the code they test
-- Easier to find and update tests when code changes
-- Clear separation between unit and integration tests
-
-### **Better Test Organization**
-- Domain-specific testing structure
-- Granular testing of specific layers
-- Clear test hierarchy and relationships
-
-### **Enhanced Developer Experience**
-- Faster test discovery
-- Better IDE support for test navigation
-- Clearer test responsibilities
-
-### **Scalability**
-- Easy to add new domains with their own integration tests
-- Consistent structure across all domains
-- Maintainable as the codebase grows
-
-## 🧪 Testing
-
-### **Running Integration Tests**
-
-```bash
-# Run all integration tests
-make test-integration
-
-# Run specific domain integration tests
-ENVIRONMENT=test INTEGRATION_TEST=true go test -v -tags=integration ./internal/ping/...
-
-# Run with test environment
-./scripts/integration_test_runner.sh
-```
-
-### **Test Results**
-
-All integration tests pass when run in the proper environment:
-- ✅ Module integration tests
-- ✅ Repository integration tests  
-- ✅ Service integration tests
-- ✅ Handler integration tests
-
-## 📚 Documentation
-
-Updated `docs/INTEGRATION_TESTING.md` includes:
-- New test structure explanation
-- Updated writing guidelines
-- Domain-specific testing examples
-- Best practices for the new structure
-
-## 🔄 Migration Notes
-
-- All existing integration test functionality is preserved
-- Test behavior remains the same
-- Build tags and environment variables unchanged
-- CI/CD integration remains compatible
-
-## 🎯 Future Considerations
-
-This structure makes it easy to:
-- Add new domains with their own integration tests
-- Implement domain-specific test utilities
-- Add performance benchmarks per domain
-- Create domain-specific test fixtures
-
-## ✅ Checklist
-
-- [x] Integration tests moved to domain-specific locations
-- [x] All test functionality preserved
-- [x] Scripts and Makefile updated
-- [x] Documentation updated
-- [x] Tests pass in proper environment
-- [x] Old integration_tests directory removed
-- [x] Code follows project conventions
-- [x] No breaking changes to existing functionality
-
-## 🔗 Related Issues
-
-This refactor addresses the need for better test organization and maintainability in the codebase.
-
----
-
-**Branch:** `refactor/integration-tests-domain-specific`  
-**Target:** `main`  
-**Type:** Refactor (non-breaking)
