--- internal/ping/adapter/ping_repo/repository_integration_test.go
@@ -1,167 +1,0 @@
-//go:build integration
-// +build integration
-
-package pingrepo_test
-
-import (
-	"context"
-	pingrepo "go-skeleton/internal/ping/adapter/ping_repo"
-	"go-skeleton/internal/ping/core/domain"
-	"os"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-)
-
-// TestPingRepositoryIntegration tests the ping repository with real database and cache
-func TestPingRepositoryIntegration(t *testing.T) {
-	// Skip if not running integration tests
-	if os.Getenv("INTEGRATION_TEST") != "true" {
-		t.Skip("Skipping integration test. Set INTEGRATION_TEST=true to run")
-	}
-
-	t.Run("Ping Repository Basic Functionality", func(t *testing.T) {
-		ctx := context.Background()
-
-		// Create repository with nil dependencies (no external calls)
-		repo := pingrepo.NewPingRepository(nil, nil)
-		var result domain.Ping
-
-		// Act
-		err := repo.Ping(ctx, &result)
-
-		// Assert
-		assert.NoError(t, err)
-		assert.Equal(t, "ping from repository", result.Message)
-	})
-
-	t.Run("Ping Repository Performance", func(t *testing.T) {
-		ctx := context.Background()
-		repo := pingrepo.NewPingRepository(nil, nil)
-		var result domain.Ping
-
-		// Test performance
-		start := time.Now()
-		err := repo.Ping(ctx, &result)
-		duration := time.Since(start)
-
-		assert.NoError(t, err)
-		assert.Less(t, duration, 50*time.Millisecond, "Repository should respond quickly")
-
-		t.Logf("Repository response time: %v", duration)
-	})
-
-	t.Run("Ping Repository Context Handling", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		repo := pingrepo.NewPingRepository(nil, nil)
-		var result domain.Ping
-
-		// Act
-		err := repo.Ping(ctx, &result)
-
-		// Assert
-		assert.NoError(t, err)
-		assert.Equal(t, "ping from repository", result.Message)
-	})
-}
-
-// TestPingRepositoryWithDatabase tests the repository with real database connection
-func TestPingRepositoryWithDatabase(t *testing.T) {
-	// Skip if not running integration tests
-	if os.Getenv("INTEGRATION_TEST") != "true" {
-		t.Skip("Skipping integration test. Set INTEGRATION_TEST=true to run")
-	}
-
-	t.Run("Ping Repository with Database", func(t *testing.T) {
-		// This test would require a real database connection
-		// For now, we'll skip it and add a placeholder
-		t.Skip("Database integration test - requires real database setup")
-
-		// TODO: Implement with real database connection
-		// 1. Set up test database connection using the test environment
-		// 2. Create repository with real DB
-		// 3. Test repository with real database
-		// 4. Verify database connectivity and ping functionality
-		// 5. Test database error scenarios
-	})
-}
-
-// TestPingRepositoryWithCache tests the repository with real cache connection
-func TestPingRepositoryWithCache(t *testing.T) {
-	// Skip if not running integration tests
-	if os.Getenv("INTEGRATION_TEST") != "true" {
-		t.Skip("Skipping integration test. Set INTEGRATION_TEST=true to run")
-	}
-
-	t.Run("Ping Repository with Cache", func(t *testing.T) {
-		// This test would require a real cache connection
-		// For now, we'll skip it and add a placeholder
-		t.Skip("Cache integration test - requires real cache setup")
-
-		// TODO: Implement with real cache connection
-		// 1. Set up test cache connection using the test environment
-		// 2. Create repository with real cache
-		// 3. Test repository with real cache
-		// 4. Verify cache connectivity and ping functionality
-		// 5. Test cache error scenarios
-	})
-}
-
-// TestPingRepositoryWithFullStack tests the repository with both database and cache
-func TestPingRepositoryWithFullStack(t *testing.T) {
-	// Skip if not running integration tests
-	if os.Getenv("INTEGRATION_TEST") != "true" {
-		t.Skip("Skipping integration test. Set INTEGRATION_TEST=true to run")
-	}
-
-	t.Run("Ping Repository with Full Stack", func(t *testing.T) {
-		// This test would require both real database and cache connections
-		// For now, we'll skip it and add a placeholder
-		t.Skip("Full stack integration test - requires real database and cache setup")
-
-		// TODO: Implement with real database and cache connections
-		// 1. Set up test database and cache connections
-		// 2. Create repository with real dependencies
-		// 3. Test repository with full stack
-		// 4. Verify both database and cache connectivity
-		// 5. Test error scenarios for both dependencies
-	})
-}
-
-// TestPingRepositoryErrorScenarios tests error handling scenarios
-func TestPingRepositoryErrorScenarios(t *testing.T) {
-	// Skip if not running integration tests
-	if os.Getenv("INTEGRATION_TEST") != "true" {
-		t.Skip("Skipping integration test. Set INTEGRATION_TEST=true to run")
-	}
-
-	t.Run("Ping Repository with Invalid Context", func(t *testing.T) {
-		// Test with cancelled context
-		ctx, cancel := context.WithCancel(context.Background())
-		cancel() // Cancel immediately
-
-		repo := pingrepo.NewPingRepository(nil, nil)
-		var result domain.Ping
-
-		// Act
-		err := repo.Ping(ctx, &result)
-
-		// Assert - should still work since we're not making external calls
-		assert.NoError(t, err)
-		assert.Equal(t, "ping from repository", result.Message)
-	})
-
-	t.Run("Ping Repository with Nil Result", func(t *testing.T) {
-		ctx := context.Background()
-		repo := pingrepo.NewPingRepository(nil, nil)
-
-		// Act - this should panic due to nil pointer dereference
-		// This is expected behavior since the code doesn't handle nil pointers
-		assert.Panics(t, func() {
-			repo.Ping(ctx, nil)
-		})
-	})
-}
