--- internal/ping/core/service/service_integration_test.go
@@ -1,186 +1,0 @@
-//go:build integration
-// +build integration
-
-package service_test
-
-import (
-	"context"
-	pingrepo "go-skeleton/internal/ping/adapter/ping_repo"
-	"go-skeleton/internal/ping/core/domain"
-	"go-skeleton/internal/ping/core/port"
-	"go-skeleton/internal/ping/core/service"
-	"os"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-)
-
-// TestPingServiceIntegration tests the ping service with real dependencies
-func TestPingServiceIntegration(t *testing.T) {
-	// Skip if not running integration tests
-	if os.Getenv("INTEGRATION_TEST") != "true" {
-		t.Skip("Skipping integration test. Set INTEGRATION_TEST=true to run")
-	}
-
-	t.Run("Ping Service with Real Dependencies", func(t *testing.T) {
-		// This test would require setting up real database and cache connections
-		// For now, we'll test the service with nil dependencies (which should still work)
-
-		ctx := context.Background()
-
-		// Create service with nil dependencies (no external calls)
-		pingRepo := pingrepo.NewPingRepository(nil, nil)
-		svcCtx := port.NewServiceContext(&pingRepo)
-		svc := service.NewPingService(&svcCtx)
-
-		var result domain.Ping
-
-		// Act
-		err := svc.Ping(ctx, &result)
-
-		// Assert
-		assert.NoError(t, err)
-		assert.Equal(t, "ping from repository", result.Message)
-	})
-
-	t.Run("Ping Service Performance", func(t *testing.T) {
-		ctx := context.Background()
-
-		pingRepo := pingrepo.NewPingRepository(nil, nil)
-		svcCtx := port.NewServiceContext(&pingRepo)
-		svc := service.NewPingService(&svcCtx)
-
-		var result domain.Ping
-
-		// Test performance
-		start := time.Now()
-		err := svc.Ping(ctx, &result)
-		duration := time.Since(start)
-
-		assert.NoError(t, err)
-		assert.Less(t, duration, 100*time.Millisecond, "Service should respond quickly")
-
-		t.Logf("Service response time: %v", duration)
-	})
-
-	t.Run("Ping Service Context Handling", func(t *testing.T) {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		pingRepo := pingrepo.NewPingRepository(nil, nil)
-		svcCtx := port.NewServiceContext(&pingRepo)
-		svc := service.NewPingService(&svcCtx)
-
-		var result domain.Ping
-
-		// Act
-		err := svc.Ping(ctx, &result)
-
-		// Assert
-		assert.NoError(t, err)
-		assert.Equal(t, "ping from repository", result.Message)
-	})
-}
-
-// TestPingServiceWithDatabase tests the service with real database connection
-func TestPingServiceWithDatabase(t *testing.T) {
-	// Skip if not running integration tests
-	if os.Getenv("INTEGRATION_TEST") != "true" {
-		t.Skip("Skipping integration test. Set INTEGRATION_TEST=true to run")
-	}
-
-	t.Run("Ping Service with Database", func(t *testing.T) {
-		// This test would require a real database connection
-		// For now, we'll skip it and add a placeholder
-		t.Skip("Database integration test - requires real database setup")
-
-		// TODO: Implement with real database connection
-		// 1. Set up test database connection
-		// 2. Create repository with real DB
-		// 3. Test service with real dependencies
-		// 4. Verify database connectivity
-	})
-}
-
-// TestPingServiceWithCache tests the service with real cache connection
-func TestPingServiceWithCache(t *testing.T) {
-	// Skip if not running integration tests
-	if os.Getenv("INTEGRATION_TEST") != "true" {
-		t.Skip("Skipping integration test. Set INTEGRATION_TEST=true to run")
-	}
-
-	t.Run("Ping Service with Cache", func(t *testing.T) {
-		// This test would require a real cache connection
-		// For now, we'll skip it and add a placeholder
-		t.Skip("Cache integration test - requires real cache setup")
-
-		// TODO: Implement with real cache connection
-		// 1. Set up test cache connection
-		// 2. Create repository with real cache
-		// 3. Test service with real dependencies
-		// 4. Verify cache connectivity
-	})
-}
-
-// TestPingServiceWithFullStack tests the service with both database and cache
-func TestPingServiceWithFullStack(t *testing.T) {
-	// Skip if not running integration tests
-	if os.Getenv("INTEGRATION_TEST") != "true" {
-		t.Skip("Skipping integration test. Set INTEGRATION_TEST=true to run")
-	}
-
-	t.Run("Ping Service with Full Stack", func(t *testing.T) {
-		// This test would require both real database and cache connections
-		// For now, we'll skip it and add a placeholder
-		t.Skip("Full stack integration test - requires real database and cache setup")
-
-		// TODO: Implement with real database and cache connections
-		// 1. Set up test database and cache connections
-		// 2. Create repository with real dependencies
-		// 3. Test service with full stack
-		// 4. Verify both database and cache connectivity
-		// 5. Test error scenarios for both dependencies
-	})
-}
-
-// TestPingServiceErrorScenarios tests error handling scenarios
-func TestPingServiceErrorScenarios(t *testing.T) {
-	// Skip if not running integration tests
-	if os.Getenv("INTEGRATION_TEST") != "true" {
-		t.Skip("Skipping integration test. Set INTEGRATION_TEST=true to run")
-	}
-
-	t.Run("Ping Service with Invalid Context", func(t *testing.T) {
-		// Test with cancelled context
-		ctx, cancel := context.WithCancel(context.Background())
-		cancel() // Cancel immediately
-
-		pingRepo := pingrepo.NewPingRepository(nil, nil)
-		svcCtx := port.NewServiceContext(&pingRepo)
-		svc := service.NewPingService(&svcCtx)
-
-		var result domain.Ping
-
-		// Act
-		err := svc.Ping(ctx, &result)
-
-		// Assert - should still work since we're not making external calls
-		assert.NoError(t, err)
-		assert.Equal(t, "ping from repository", result.Message)
-	})
-
-	t.Run("Ping Service with Nil Result", func(t *testing.T) {
-		ctx := context.Background()
-
-		pingRepo := pingrepo.NewPingRepository(nil, nil)
-		svcCtx := port.NewServiceContext(&pingRepo)
-		svc := service.NewPingService(&svcCtx)
-
-		// Act - this should panic due to nil pointer dereference
-		// This is expected behavior since the code doesn't handle nil pointers
-		assert.Panics(t, func() {
-			svc.Ping(ctx, nil)
-		})
-	})
-}
